Renamed libs.API files to only lowercase
