+++ Oanda/libs/API/__init__.py
@@ -5,7 +5,7 @@ root_path = os.path.relpath(__file__+'/../../../')
 credentials_path = os.path.relpath(root_path + '/configs/credentials/credentials.yaml')
 with open(credentials_path, "r") as cfgfile:
     API_CONFIG = yaml.safe_load(cfgfile)
-from .Oanda import *
-# from .Order_Position_Book import *
-# from .Orders import *
-from .WorkingFunctions import readable_output
+from .oanda import *
+# from .order_position_book import *
+# from .orders import *
+from .working_functions import readable_output
